Improve HUD
Using the assets from Squadron Dragon, I made a massive overhaul with the
HUD. Logic is a bit basic but should do for now. Shooting to be
implemented. HUD is now a prefab. Known bug: adding HUD prefab to seen
will offset the button and text in the pause menu. Must either fix this
bug for fix it manually everytime you use HUD.